--- src/main/java/com/riwiEcomerce/riwiEcomerce/domain/entities/UserEntity.java
@@ -1,5 +1,0 @@
-package com.riwiEcomerce.riwiEcomerce.domain.entities;
-
-public class UserEntity {
-    
-}
