+++ src/main/java/com/riwi_ecomerce/riwi_ecomerce/domain/entities/UserEntity.java
@@ -1,0 +1,5 @@
+package com.riwi_ecomerce.riwi_ecomerce.domain.entities;
+
+public class UserEntity {
+    
+}
